+++ client/src/components/charts/bar/BarChart.js
@@ -1,7 +1,7 @@
-import React, { useEffect } from "react";
-import PropTypes from "prop-types";
-import * as d3 from "d3";
-import BaseChart from "./BaseChart";
+import React, { useEffect } from 'react';
+import PropTypes from 'prop-types';
+import * as d3 from 'd3';
+import BaseChart from '../BaseChart';
 
 /*
  * @param data = [{x: 2, y: 2}, ...]
