+++ client/src/components/charts/bar/StackedBarChart.js
@@ -1,9 +1,9 @@
-import React, { useEffect } from "react";
-import PropTypes from "prop-types";
-import * as d3 from "d3";
-import BaseChart from "./BaseChart";
+import React, { useEffect } from 'react';
+import PropTypes from 'prop-types';
+import * as d3 from 'd3';
+import BaseChart from '../BaseChart';
 
-class BandedBarChart extends BaseChart {
+class StackedBarChart extends BaseChart {
   constructor(options) {
     super(options);
 
@@ -47,11 +47,6 @@ class BandedBarChart extends BaseChart {
       .scaleBand()
       .domain(data.map((d) => d.x))
       .range([this.margins.left, this.dimensions.innerWidth]);
-
-    this.scaleX2 = d3
-      .scaleBand()
-      .domain(new Set([...data.map((d) => d.x2)]))
-      .range([0, this.scaleX.bandwidth() - this.margins.right]);
 
     this.scaleY = d3
       .scaleLinear()
@@ -125,12 +120,9 @@ class BandedBarChart extends BaseChart {
       .enter()
       .append("rect")
       .attr("class", "rects")
-      .attr(
-        "x",
-        (d) => this.scaleX(d.x) + this.scaleX2(d.x2) - this.margins.left / 2
-      )
+      .attr("x", (d) => this.scaleX(d.x))
       .style("fill", (d) => this.colorFxn(d.x2))
-      .attr("width", this.scaleX2.bandwidth())
+      .attr("width", this.scaleX.bandwidth() - this.margins.left * 2)
       .attr("height", 0)
       .attr("y", this.scaleY(0));
 
@@ -156,11 +148,6 @@ class BandedBarChart extends BaseChart {
       .scaleLinear()
       .domain([0, d3.max(data, (d) => d.y) + 1])
       .range([this.margins.left, this.dimensions.innerWidth]);
-
-    this.scaleX2 = d3
-      .scaleBand()
-      .domain(new Set([...data.map((d) => d.x2)]))
-      .range([0, this.scaleX.bandwidth() - this.margins.right]);
 
     this.axisX = d3.axisLeft(this.scaleX).ticks(data.map((d) => d.x).length);
     this.axisY = d3.axisTop(this.scaleY).ticks(5);
@@ -221,13 +208,10 @@ class BandedBarChart extends BaseChart {
       .data(data)
       .enter()
       .append("rect")
-      .attr(
-        "y",
-        (d) => this.scaleX(d.x) + this.scaleX2(d.x2) - this.margins.top / 2
-      )
+      .attr("y", (d) => this.scaleX(d.x) - this.margins.top / 2)
       .style("fill", (d) => this.colorFxn(d.x2))
       .attr("width", 0)
-      .attr("height", this.scaleX2.bandwidth())
+      .attr("height", this.scaleX.bandwidth() - this.margins.left)
       .attr("x", this.scaleY(0));
 
     this.bars
@@ -250,7 +234,7 @@ class BandedBarChart extends BaseChart {
   }
 }
 
-const BandedBarComponent = (props) => {
+const StackedBarComponent = (props) => {
   const { data, options } = props;
 
   const defaultOptions = {
@@ -268,20 +252,23 @@ const BandedBarComponent = (props) => {
     }
   };
 
+  /* values should be sorted so longer rectangles are rendered first */
+  let sortedData = [...data].sort((a, b) => b.y - a.y);
+
   const getOptions = setOptions(options);
 
   useEffect(() => {
-    const chart = new BandedBarChart(getOptions);
+    const chart = new StackedBarChart(getOptions);
     chart.createChart();
-    chart.displayData(getOptions, data);
+    chart.displayData(getOptions, sortedData);
   }, [options]);
 
   return <div id={getOptions.containerId} className="chart-viewbox"></div>;
 };
 
-BandedBarComponent.propTypes = {
+StackedBarComponent.propTypes = {
   data: PropTypes.any,
   options: PropTypes.any,
 };
 
-export default BandedBarComponent;
+export default StackedBarComponent;
