+++ client/src/components/charts/bar/index.js
@@ -1,0 +1,9 @@
+import BarChartComponent from './BarChart';
+import BandedBarComponent from './BandedBarChart';
+import StackedBarComponent from './StackedBarChart';
+
+export {
+	BarChartComponent,
+	BandedBarComponent,
+	StackedBarComponent
+};
