+++ client/src/routes/index.js
@@ -70,7 +70,22 @@ const cardRoutes = [
             type: "page",
             level: 3,
           },
-        ],
+          {
+            name: "line_chart",
+            title: "Line Chart",
+            component: ChartPage,
+            chart: 'line-chart',
+            image: {
+              src: barCharts,
+              alt: "Placeholder",
+            },
+            description:
+              "An interactive, customizable line chart",
+            path: "/data-viz/chart-card/line-chart",
+            type: "page",
+            level: 3,
+          }          
+        ]
       },
       {
         name: "tableau",
